+++ src/pages/SourceListPage.tsx
@@ -5,9 +5,9 @@ import { Link } from "react-router";
 
 import { Button } from "react-toolbox/lib/button";
 
-import { List } from "../components/List/List";
-import TwoLineItem from "../components/List/TwoLineItem";
-import { TwoPane } from "../components/TwoPane";
+import List from "../components/List/List";
+import ListItem from "../components/List/ListItem";
+import TwoPane from "../components/TwoPane";
 
 import Source from "../models/source";
 
@@ -77,12 +77,11 @@ class SourceList extends React.Component<SourceListProps, SourceListState> {
     renderItem(index: number, key: string): JSX.Element {
         let source = this.props.sources[index];
         return (
-            // <Link to={"/skills/" + source.id} style={{ textDecoration: "none" }}>
-                <TwoLineItem
-                    index={index}
-                    primaryValue={source.name}
-                    secondaryValue={moment(source.created).format("MMM Do, YYYY")} />
-            // </Link>
+            <ListItem
+                index={index}
+                primaryValue={source.name}
+                to={"/skills/" + source.id}
+                secondaryValue={moment(source.created).format("MMM Do, YYYY")} />
         );
     }
